+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Language
 - [feat (#21): [syntax/scanner, syntax/parser, typechecker, codegen] implement struct types & constructors](https://github.com/amp-lang/amp/issues/21)
 - [feat (#23): [cli] implement multiple module linking](https://github.com/amp-lang/amp/issues/23)
+- [feat (#22): [syntax/scanner, syntax/parser, typechecker, codegen] implement struct field accessing](https://github.com/amp-lang/amp/issues/22)
 
 ## v0.2.0-alpha